Deconvolution

DeconvOptim.deconvolution — Function

deconvolution(measured, psf; <keyword arguments>)

Computes the deconvolution of measured and psf. Return parameter is a tuple with two elements. The first entry is the deconvolved image. The second return parameter is the output of the optimization of Optim.jl

Multiple keyword arguments can be specified for different loss functions, regularizers and mappings.

Arguments

loss=Poisson(): the loss function taking a vector the same shape as measured.

regularizer=GR(): A regularizer function, same form as loss.

λ=0.05: A float indicating the total weighting of the regularizer with respect to the global loss function

background=0: A float indicating a background intensity level.

mapping=Non_negative(): Applies a mapping of the optimizer weight. Default is a parabola which achieves a non-negativity constraint.

iterations=20: Specifies a number of iterations after the optimization. definitely should stop.

conv_dims: A tuple indicating over which dimensions the convolution should happen. per default conv_dims=1:ndims(psf)

plan_fft=true: Boolean whether plan_fft is used. Gives a slight speed improvement.

padding=0: an float indicating the amount (fraction of the size in that dimension) of padded regions around the reconstruction. Prevents wrap around effects of the FFT. A array with size(arr)=(400, 400) with padding=0.05 would result in reconstruction size of (440, 440). However, we only return the reconstruction cropped to the original size. padding=0 disables any padding.

optim_options=nothing: Can be a options file required by Optim.jl. Will overwrite iterations.

optim_optimizer=LBFGS(): The chosen Optim.jl optimizer.

Example

julia> using DeconvOptim, TestImages, Colors, Noise;

julia> img = Float32.(testimage("resolution_test_512"));

julia> psf = Float32.(generate_psf(size(img), 30));

julia> img_b = conv_psf(img, psf);

julia> img_n = poisson(img_b, 300);

julia> @time res, o = deconvolution(img_n, psf);

DeconvOptim.richardson_lucy_iterative — Function

richardson_lucy_iterative(measured, psf; <keyword arguments>)

Classical iterative Richardson-Lucy iteration scheme for deconvolution. measured is the measured array and psf the point spread function. Converges slower than the optimization approach of deconvolution

Keyword Arguments

regularizer=GR(): A regularizer function. Can be exchanged

λ=0.05: A float indicating the total weighting of the regularizer with respect to the global loss function

iterations=100: Specifies number of iterations.

Example

julia> using DeconvOptim, TestImages, Colors, Noise;

julia> img = Float32.(testimage("resolution_test_512"));

julia> psf = Float32.(generate_psf(size(img), 30));

julia> img_b = conv_psf(img, psf);

julia> img_n = poisson(img_b, 300);

julia> @time res = richardson_lucy_iterative(img_n, psf);

More generic alternative

DeconvOptim.invert — Function

invert(measured, rec0, forward; <keyword arguments>)

Tries to invert the forward model. forward is a function taking an input with the shape of rec0 and returns an object which has the same shape as measured

Multiple keyword arguments can be specified for different loss functions, regularizers and mappings.

Arguments

loss=Poisson(): the loss function being compatible to compare with measured.

regularizer=nothing: A regularizer function, same form as loss.

λ=0.05: A float indicating the total weighting of the regularizer with respect to the global loss function

mapping=Non_negative(): Applies a mapping of the optimizer weight. Default is a parabola which achieves a non-negativity constraint.

iterations=10: Specifies a number of iterations after the optimization. definitely should stop. Will be overwritten if optim_options is provided.

optim_options=nothing: Can be a options file required by Optim.jl. Will overwrite iterations.

optim_optimizer=LBFGS(): The chosen Optim.jl optimizer.