Loss Functions

DeconvOptim.Poisson — Function

Poisson()

Returns a function to calculate Poisson loss Check the help of poisson_aux.

DeconvOptim.poisson_aux — Function

poisson_aux(μ, meas, storage=similar(μ))

Calculates the Poisson loss for μ and meas. μ can be of larger size than meas. In that case we extract a centered region from μ of the same size as meas.

DeconvOptim.Gauss — Function

Gauss()

Returns a function to calculate Gauss loss. Check the help of gauss_aux.

DeconvOptim.gauss_aux — Function

gauss_aux(μ, meas, storage=similar(μ))

Calculates the Gauss loss for μ and meas. μ can be of larger size than meas. In that case we extract a centered region from μ of the same size as meas.

DeconvOptim.ScaledGauss — Function

ScaledGauss()

Returns a function to calculate scaled Gauss loss. Check the help of scaled_gauss_aux.

DeconvOptim.scaled_gauss_aux — Function

scaled_gauss_aux(μ, meas, storage=similar(μ); read_var=0)

Calculates the scaled Gauss loss for μ and meas. read_var=0 is the readout noise variance of the sensor. μ can be of larger size than meas. In that case we extract a centered region from μ of the same size as meas.