Mapping Functions

DeconvOptim.Non_negative — Function

Non_negative()

Returns a function and an inverse function inverse function to map numbers to non-negative numbers. We use a parabola.

Examples

julia> p, p_inv = Non_negative()
(DeconvOptim.var"#5#7"(), DeconvOptim.var"#6#8"())

julia> x = [-1, 2, -3]
3-element Array{Int64,1}:
 -1
  2
 -3

julia> p(x)
3-element Array{Int64,1}:
 1
 4
 9

julia> p_inv(p(x))
3-element Array{Float64,1}:
 1.0
 2.0
 3.0

DeconvOptim.Map_0_1 — Function

Map_0_1()

Returns a function and an inverse function to map numbers to an interval between 0 and 1. via an exponential function.

DeconvOptim.Piecewise_positive — Function

Piecewise_positive()

Returns a function and an inverse function to map numbers to larger than 0 via two function stitched together.

DeconvOptim.Pow4_positive — Function

Pow4_positive()

Returns a function and an inverse function to map numbers to larger than 0 with abs2.(abs2.(x))

DeconvOptim.Abs_positive — Function

Abs_positive()

Returns a function and an inverse function to map numbers to larger than 0.