Regularizers

DeconvOptim.TV — Function

TV(; <keyword arguments>)

This function returns a function to calculate the Total Variation regularizer of a n-dimensional array.

Arguments

num_dim=2:

sum_dims=[1, 2, 3]: A array containing the dimensions we want to sum over

weights=[1, 1, 0.25]: A array containing weights to weight the contribution of different dimensions

step=1: A integer indicating the step width for the array indexing

mode="central": Either "central" or "forward" accounting for different modes of the spatial gradient. Default is "central".

Examples

To create a regularizer for a 3D dataset where the third dimension has different contribution. For the derivative we use forward mode.

julia> TV(num_dim=3, sum_dims=[1, 2, 3], weights=[1, 1, 0.25], mode="forward")

DeconvOptim.Tikhonov — Function

Tikhonov(; <keyword arguments>)

This function returns a function to calculate the Tikhonov regularizer of a n-dimensional array.

Arguments

num_dim=2:

sum_dims=[1, 2]: A array containing the dimensions we want to sum over

weights=[1, 1]: A array containing weights to weight the contribution of different dimensions

step=1: A integer indicating the step width for the array indexing

mode="laplace": Either "laplace", "spatial_grad_square", "identity" accounting for different modes of the Tikhonov regularizer. Default is "laplace".

Examples

To create a regularizer for a 3D dataset where the third dimension has different contribution.

julia> Tikhonov(sum_dims=[1, 2, 3], weights=[1, 1, 0.25])

DeconvOptim.GR — Function

GR(; <keyword arguments>)

This function returns a function to calculate the Good's roughness regularizer of a n-dimensional array.

Arguments

num_dim=2: Dimension of the array that should be regularized

sum_dims=[1, 2]: A array containing the dimensions we want to sum over

weights=[1, 1]: A array containing weights to weight the contribution of different dimensions

step=1: A integer indicating the step width for the array indexing

mode="forward": Either "central" or "forward" accounting for different modes of the spatial gradient. Default is "central".

Examples

To create a regularizer for a 3D dataset where the third dimension has different contribution. For the derivative we use forward mode.

julia> GR(sum_dims=[1, 2, 3], weights=[1, 1, 0.25], mode="forward")