Util functions

Convolution Functions

DeconvOptim.conv_psf — Function

conv_psf(obj, psf [, dims])

Convolve obj with psf over dims dimensions. Based on FFT convolution.

DeconvOptim.conv_otf — Function

conv_otf(obj, otf [, dims])

Performs a FFT-based convolution of an obj with an otf. otf = fft(psf). The 0 frequency of the otf must be located at position [1, 1, 1]. The obj can be of arbitrary dimension but ndims(psf) ≥ ndims(otf). The convolution happens over the dims array. Any further dimensions are broadcasted. Per default dims = [1, 2].

DeconvOptim.conv_otf_r — Function

conv_otf_r(obj, otf [, dims])

Performs a FFT-based convolution of an obj with an otf. Same arguments as conv_otf but with obj being real and otf=rfft(psf).

DeconvOptim.plan_conv_r — Function

plan_conv_r(psf [, dims])

Pre-plan an optimized convolution for array shaped like psf (based on pre-plan FFT) along the given dimenions dims. dims = [1, 2] per default. The 0 frequency of psf must be located at [1, 1, 1]. We return first the otf (obtained by rfft(psf)). The second return is the convolution function conv. conv itself has two arguments. conv(obj, otf) where obj is the object and otf the otf.

This function achieves faster convolution than conv_psf(obj, psf).

Point Spread Function

DeconvOptim.generate_psf — Function

generate_psf(psf_size, radius)

Generation of an approximate 2D PSF. psf_size is the output size of the PSF. The PSF will be centered around the point [1, 1], radius indicates the pupil diameter in pixel from which the PSF is generated.

Examples

julia> generate_psf([5, 5], 2)
5×5 Array{Float64,2}:
 0.36       0.104721    0.0152786    0.0152786    0.104721
 0.104721   0.0304627   0.00444444   0.00444444   0.0304627
 0.0152786  0.00444444  0.000648436  0.000648436  0.00444444
 0.0152786  0.00444444  0.000648436  0.000648436  0.00444444
 0.104721   0.0304627   0.00444444   0.00444444   0.0304627

Interpolation and downsampling

DeconvOptim.generate_downsample — Function

generate_downsample(num_dim, downsample_dims, factor)

Generate a function (based on Tullio.jl) which can be used to downsample arrays. num_dim (Integer) are the dimensions of the array. downsample_dims is a list of which dimensions should be downsampled. factor is a downsampling factor. It needs to be an integer number.

Examples

julia> ds = generate_downsample(2, [1, 2], 2) 
[...]
julia> ds([1 2; 3 4; 5 6; 7 8])
2×1 Array{Float64,2}:
 2.5
 6.5

julia> ds = generate_downsample(2, [1], 2)
[...]
julia> ds([1 2; 3 5; 5 6; 7 8])
2×2 Array{Float64,2}:
 2.0  3.5
 6.0  7.0

DeconvOptim.my_interpolate — Function

my_interpolate(arr, size_n, [interp_type])

Interpolates arr to the sizes provided in size_n. Therefore it holds ndims(arr) == length(size_n). interp_type specifies the interpolation type. See Interpolations.jl for all options

Center Methods

DeconvOptim.center_extract — Function

center_extract(arr, new_size)

Extracts a center of an array. new_size must be list of sizes indicating the output size of each dimension. Centered means that a center frequency stays at the center position. Works for even and uneven. If length(new_size) < length(size(arr)) the remaining dimensions are untouched and copied.

Examples

julia> center_extract([[1,2] [3, 4]], [1])
1×2 Array{Int64,2}:
 2  4

julia> center_extract([[1,2] [3, 4]], [1, 1])
1×1 Array{Int64,2}:
4

DeconvOptim.center_set! — Function

center_set!(arr_large, arr_small)

Puts the arr_small central into arr_large. The convention, where the center is, is the same as the definition as for FFT based centered. Function works both for even and uneven arrays.

Examples

julia> center_set!([1, 1, 1, 1, 1, 1], [5, 5, 5])
6-element Array{Int64,1}:
 1
 1
 5
 5
 5
 1

DeconvOptim.center_pos — Function

center_pos(x)

Calculate the position of the center frequency. Size of the array is x

Examples

julia> center_pos(3)
2
julia> center_pos(4)
4